requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
Django==3.2.6
django-xhtml2pdf==0.0.4
djangorestframework==3.12.4
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
mysqlclient==2.0.3
Pillow==8.3.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.4
pytz==2021.1
reportlab==3.6.1
six==1.16.0
sqlparse==0.4.1
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
